Add plant site upload table
Why: makes it a little simpler to handle what data needs to be uploaded,
and also the id for when the plant is offline is an autoincrementing
number as opposed to a uuid. I think it is better to consider them as
different types and let the plant sites from the server have their own
type with any extra properties that are created as a result of putting
them in the database.