Fix issue with rendering of plant sites uploads that are deleted
Why: I'm not sure if it is how the shadow dom works or not, but the
component needs a useEffect that is listening to its properties, in
order to render correctly; it must be using the components that are
already there in the dom and just overwriting the properties. Weird.