Map dom markers to canvas position
Had to take out some logic that made zooming in the map nicer (it was
adding a pan to the zoom), but the marker elements are lined up quite
well. Strange things with rendering though; dom renders faster than the
canvas, but if there is a hot reload then they line up perfectly.