Reduce load on plant list view.
Problem is that the plants site upload photos are in the same model as
the plant site upload. This makes it super heavy in the view when you
have many plant site uploads in the list as it crashes due to lack of
memory.

Best idea would be to separate out the plant site upload table again and
only load them when needed.