Add generics to search filter class to search subtypes of PlantSite
Why: Given that the closest plants page needs a model with an added
distance property on it, figured out that you can enforce the
general shape of a generic to pass into the search plant sites filter,
which means we can search plant sites of any subclass of PlantSite, and
have it return the correct data.
Also can add a default type, which will mean we will only need the
speecify the type in special cases.