Add login service to log user in
With credentials must present and set to 'include' as we are technically
sending to another domain. CORS will only allow this header on the
frontend domains url.

Also, it looks like the cookie is not set in the dev tools, but it is
and it is hidden. You can click on the cookies tab in the request
object to see that it is being sent