Switch out from using preloading to live query updates on plants page
Why: This is to improve the UX on onboarding; currently the plants page
will be an empty page if the data hasn't been downloaded before
navigating to it. Only issue I can oversee is the overhead of too many
observables, or the ui rendering too many times as plants are added.
Debouncing the state update could be an option.